dedicated unit tests for CLI and __init__, code style & format
