+++ py2puml/__main__.py
@@ -1,4 +1,4 @@
 from py2puml.cli import run
 
-if __name__ == "__main__":
+if __name__ == '__main__':
     run()
